refactor: Create project module move file.
