+++ src/app/project/create/Form/FormStepper.tsx
@@ -3,7 +3,7 @@ import StepConnector, {
   stepConnectorClasses,
 } from "@mui/material/StepConnector";
 import { styled } from "@mui/material/styles";
-import { useFormControlContext } from "@/contexts/FormControlContext";
+import { useFormControlContext } from "@/app/project/create/FormControlContext";
 
 const QontoConnector = styled(StepConnector)(({ theme }) => ({
   [`&.${stepConnectorClasses.alternativeLabel}`]: {
